move google services into book service
